refactor(frontend/services): rename service modules to kebab-case
- guardianService.js → guardian-service.js (updated imports to core/api/client)
- feeService.js → fee-service.js
- documentService.js → document-service.js
- mockDataService.js → mock-data-service.js
- studentService.js → student-service.js
- setupTests.js → setup-tests.js
- Add barrel export index.js for clean cross-feature imports
- Add services/README.md with domain naming conventions
