+++ student-board/src/services/document-service.js
@@ -1,0 +1,8 @@
+import apiClient from '../core/api/client.js';
+
+export const documentService = {
+  getDocuments: async (studentId) => {
+    const response = await apiClient.get(`/students/${studentId}/documents`);
+    return response.data.data;
+  },
+};
