--- student-board/src/services/documentService.js
@@ -1,8 +1,0 @@
-import api from './api';
-
-export const documentService = {
-  getDocuments: async (studentId) => {
-    const response = await api.get(`/students/${studentId}/documents`);
-    return response.data.data;
-  },
-};
